Update to pass input text via Intent
Unit tests were testing whether the text passed to second activity matched what is entered into the text field of first activity, but the text was never passed to second activity, so tests failed. Now MainActivity takes the text from text field and adds as an Extra to the intent used to start second activity.